+++ Source/MurderInSpace/Private/Modes/MyState.cpp
@@ -10,7 +10,7 @@
 
 DEFINE_LOG_CATEGORY(LogMyGame);
 
-float UMyState::GetInitialAngularVelocity(FRnd Rnd)
+double UMyState::GetInitialAngularVelocity(FRnd Rnd)
 {
 	return Rnd.Poisson(Rnd.RndGen) / 1.e3;
 }
